+++ src/py42/sdk/queries/query_filter.py
@@ -1,14 +1,15 @@
-# -*- coding: utf-8 -*-
+from collections import OrderedDict
 from datetime import datetime
 
 from py42._internal.compat import str
+from py42._internal.compat import string_type
 from py42.util import convert_datetime_to_timestamp_str
 from py42.util import convert_timestamp_to_str
 
 
 def create_query_filter(term, operator, value=None):
-    """Creates a :class:`QueryFilter` object. Useful for creating filters that are not yet
-    defined in py42 or programmatically crafting filters.
+    """Creates a :class:`QueryFilter` object. Useful for programmatically crafting query
+    filters, such as filters not yet defined in py42.
 
     Args:
         term (str): The term of the filter, such as ``actor`` or ``sharedWith``.
@@ -23,10 +24,10 @@ def create_query_filter(term, operator, value=None):
 
 
 def create_filter_group(query_filter_list, filter_clause):
-    """Creates a :class:`FilterGroup` object. Useful for creating custom filter groups not yet
-    supported in py42 or programmatically crafting filter groups. Alternatively, if you
-    want to create custom filter groups with previously defined operators such `IS` or `IS_IN`,
-    see the other methods in this module, such as :meth:`create_eq_filter_group()`.
+    """Creates a :class:`FilterGroup` object. Useful for programmatically crafting query
+    filters, such as filters not yet defined in py42. Alternatively, if you want to create
+    custom filter groups with previously defined operators such `IS` or `IS_IN`, see the
+    other methods in this module, such as :meth:`create_eq_filter_group()`.
 
     Args:
         query_filter_list (list): a list of :class:`QueryFilter` objects.
@@ -41,8 +42,8 @@ def create_filter_group(query_filter_list, filter_clause):
 
 def create_eq_filter_group(term, value):
     """"Creates a :class:`FilterGroup` for filtering results where the value with key ``term``
-    equals the given value. Useful for creating ``IS`` filters that are not yet supported in py42
-    or programmatically crafting filter groups.
+    equals the given value. Useful for creating ``IS`` filters that are not yet supported
+    in py42 or programmatically crafting filter groups.
 
     Args:
         term: (str): The term of the filter, such as ``actor`` or ``sharedWith``.
@@ -75,8 +76,8 @@ def create_not_eq_filter_group(term, value):
 
 def create_is_in_filter_group(term, value_list):
     """"Creates a :class:`FilterGroup` for filtering results where the value with key ``term``
-    is one of several values. Useful for creating ``IS_IN`` filters that are not yet supported in
-    py42 or programmatically crafting filter groups.
+    is one of several values. Useful for creating ``IS_IN`` filters that are not yet supported
+    in py42 or programmatically crafting filter groups.
 
     Args:
         term: (str): The term of the filter, such as ``actor`` or ``sharedWith``.
@@ -114,7 +115,7 @@ def create_on_or_after_filter_group(term, value):
     crafting filter groups.
 
     Args:
-        term: (str): The term of the filter, such as ``actor`` or ``sharedWith``.
+        term: (str): The term of the filter, such as ``eventTimestamp``.
         value (str or int): The value used to filter results.
 
     Returns:
@@ -132,7 +133,7 @@ def create_on_or_before_filter_group(term, value):
     crafting filter groups.
 
     Args:
-        term: (str): The term of the filter, such as ``actor`` or ``sharedWith``.
+        term: (str): The term of the filter, such as ``eventTimestamp``.
         value (str or int): The value used to filter results.
 
     Returns:
@@ -150,7 +151,7 @@ def create_in_range_filter_group(term, start_value, end_value):
     in py42 or programmatically crafting filter groups.
 
     Args:
-        term: (str): The term of the filter, such as ``actor`` or ``sharedWith``.
+        term: (str): The term of the filter, such as ``eventTimestamp``.
         start_value (str or int): The start value used to filter results.
         end_value (str or int): The end value used to filter results.
 
@@ -172,8 +173,8 @@ class QueryFilterStringField(object):
 
     @classmethod
     def eq(cls, value):
-        """Returns a :class:`FilterGroup` to find events where the filter equals the provided
-        ``value``.
+        """Returns a :class:`FilterGroup` that is useful for finding results where the value
+        with key ``self._term`` equals the provided ``value``.
 
         Args:
             value (str): The value to match on.
@@ -185,11 +186,11 @@ class QueryFilterStringField(object):
 
     @classmethod
     def not_eq(cls, value):
-        """Returns a :class:`FilterGroup` to find events where the filter is not equal to the
-        provided ``value``.
+        """Returns a :class:`FilterGroup` that is useful for finding results where the value
+        with key ``self._term`` does not equal the provided ``value``.
 
         Args:
-            value (str): The value to exclude file events on.
+            value (str): The value to exclude on.
 
         Returns:
             :class`FilterGroup`
@@ -198,8 +199,8 @@ class QueryFilterStringField(object):
 
     @classmethod
     def is_in(cls, value_list):
-        """Returns a :class:`FilterGroup` to find events where the filter is in the provided
-        ``value_list``.
+        """Returns a :class:`FilterGroup` that is useful for finding results where the value
+        with the key ``self._term`` is in the provided ``value_list``.
 
         Args:
             value_list (list): The list of values to match on.
@@ -211,8 +212,8 @@ class QueryFilterStringField(object):
 
     @classmethod
     def not_in(cls, value_list):
-        """Returns a :class:`FilterGroup` to find events where the filter is not in the provided
-        ``value_list``.
+        """Returns a :class:`FilterGroup` that is useful for finding results where the value
+        with the key ``self._term`` is not in the provided ``value_list``.
 
         Args:
             value_list (list): The list of values to exclude on.
@@ -230,8 +231,8 @@ class QueryFilterTimestampField(object):
 
     @classmethod
     def on_or_after(cls, value):
-        """Returns a :class:`FilterGroup` to find events where the filter timestamp is on or after
-        the provided `value`.
+        """Returns a :class:`FilterGroup` that is useful for finding results where the value
+        with key ``self._term` is on or after the provided ``value``.
 
         Args:
             value (str or int): The value used to filter results.
@@ -244,8 +245,8 @@ class QueryFilterTimestampField(object):
 
     @classmethod
     def on_or_before(cls, value):
-        """Returns a :class:`FilterGroup` to find events where the filter timestamp is on or before
-        the provided `value`.
+        """Returns a :class:`FilterGroup` that is useful for finding results where the value
+        with key ``self._term`` is on or before the provided `value`.
 
         Args:
             value (str or int): The value used to filter results.
@@ -258,8 +259,8 @@ class QueryFilterTimestampField(object):
 
     @classmethod
     def in_range(cls, start_value, end_value):
-        """Returns a :class:`FilterGroup` to find events where the filter timestamp is in range
-        between the provided `start_value` and `end_value`.
+        """Returns a :class:`FilterGroup` that is useful for finding results where the value
+        with key ``self._term`` is in range between the provided `start_value` and `end_value``.
 
         Args:
             start_value (str or int): The start value used to filter results.
@@ -276,8 +277,8 @@ class QueryFilterTimestampField(object):
 
     @classmethod
     def on_same_day(cls, value):
-        """Returns a :class:`FilterGroup` to find events where the filter timestamp is within the
-        same calendar day as the provided `value`.
+        """Returns a :class:`FilterGroup` that is useful for finding results where the value
+        with key ``self._term`` is within the same calendar day as the provided `value`.
 
         Args:
             value (str or int): The value used to filter results.
@@ -306,7 +307,8 @@ class QueryFilterBooleanField(object):
 
     @classmethod
     def is_true(cls):
-        """Returns a :class:`FilterGroup` to find events where the filter state is True.
+        """Returns a :class:`FilterGroup` that is useful for finding results where the value
+        with key ``self._term`` is True.
 
         Returns:
             :class`FilterGroup`
@@ -315,7 +317,8 @@ class QueryFilterBooleanField(object):
 
     @classmethod
     def is_false(cls):
-        """Returns a :class:`FilterGroup` to find events where the filter state is False.
+        """Returns a :class:`FilterGroup` that is useful for finding results where the value
+        with key ``self._term`` is False.
 
         Returns:
             :class`FilterGroup`
@@ -324,15 +327,16 @@ class QueryFilterBooleanField(object):
 
 
 class QueryFilter(object):
-    """Class for constructing a single filter object for use in a Forensic Search query.
+    """Class for constructing a single filter object for use in a search query.
 
     When :func:`str()` is called on a :class:`QueryFilter` instance, the (``term``, ``operator``,
-    ``value``) attribute combination is transformed into a JSON string to be used as part of a
-    Forensic Search or Alert query.
+    ``value``) attribute combination is transformed into a JSON string to be used as part
+    of a Forensic Search or Alert query.
 
     When :func:`dict()` is called on a :class:`QueryFilter` instance, the (``term``, ``operator``,
-    ``value``) attribute combination is transformed into the Python `dict` equivalent of their JSON representation. This can be useful
-    for programmatically manipulating a :class:`QueryFilter` after it's been created.
+    ``value``) attribute combination is transformed into the Python `dict` equivalent of
+    their JSON representation. This can be useful for programmatically manipulating a
+    :class:`QueryFilter` after it's been created.
     """
 
     _term = None
@@ -381,19 +385,29 @@ class QueryFilter(object):
         )
 
     def __iter__(self):
-        output_dict = {
-            u"operator": self._operator,
-            u"term": self._term,
-            u"value": self._value,
-        }
+        output_dict = OrderedDict()
+        output_dict[u"operator"] = self._operator
+        output_dict[u"term"] = self._term
+        output_dict[u"value"] = self._value
         for key in output_dict:
             yield key, output_dict[key]
 
+    def __eq__(self, other):
+        if isinstance(other, (QueryFilter, tuple, list)):
+            return tuple(self) == tuple(other)
+        elif isinstance(other, string_type):
+            return str(self) == other
+        else:
+            return False
+
+    def __hash__(self):
+        return hash(str(self))
+
 
 class FilterGroup(object):
-    """Class for constructing a logical sub-group of related filters from a list of QueryFilter
-    objects. Takes a list of QueryFilter objects and combines them logically using the passed in
-    filter clause (``AND`` or ``OR``).
+    """Class for constructing a logical sub-group of related filters from a list of
+    :class:`QueryFilter` objects. Takes a list of :class:`QueryFilter` objects and combines them
+    logically using the passed in filter clause (``AND`` or ``OR``).
 
     When :func:`str()` is called on a :class:`FilterGroup` instance, the combined filter items are
     transformed into a JSON string to be used as part of a Forensic Search or Alert query.
@@ -408,20 +422,18 @@ class FilterGroup(object):
         self._filter_clause = filter_clause
 
     @classmethod
-    def from_dict(cls, _dict, filter_clause=u"AND"):
-        """Creates an instance of :class:`FilterGroup` from the values found in ``_dict``
-        and the given clause. ``_dict`` must contain key ``filters`` and correctly represent
-        :class:`QueryFilter` JSON objects.
+    def from_dict(cls, _dict):
+        """Creates an instance of :class:`FilterGroup` from the values found in ``_dict``.
+        ``_dict`` must contain keys ``filters`` and ``filterClause``.
 
         Args:
             _dict (dict): A dictionary containing keys ``term``, ``operator``, and ``value``.
-            filter_clause (str): The clause joining the filters, such as ``AND`` or ``OR``.
 
         Returns:
             :class:`FilterGroup`
         """
         filter_list = [QueryFilter.from_dict(item) for item in _dict[u"filters"]]
-        return cls(filter_list, filter_clause=filter_clause)
+        return cls(filter_list, filter_clause=_dict[u"filterClause"])
 
     @property
     def filter_list(self):
@@ -435,16 +447,40 @@ class FilterGroup(object):
 
         return self._filter_clause
 
+    @filter_clause.setter
+    def filter_clause(self, value):
+        """The clause joining the filters, such as ``AND`` or ``OR``."""
+
+        self._filter_clause = value
+
+    @property
+    def _filter_set(self):
+        return sorted(list(set(self.filter_list)), key=str)
+
     def __str__(self):
-        filters_string = u",".join(
-            str(filter_item) for filter_item in self._filter_list
-        )
+        filters_string = u",".join(str(filter_item) for filter_item in self._filter_set)
         return u'{{"filterClause":"{0}", "filters":[{1}]}}'.format(
             self._filter_clause, filters_string
         )
 
     def __iter__(self):
-        filter_list = [dict(item) for item in self._filter_list]
+        filter_list = [dict(item) for item in self._filter_set]
         output_dict = {u"filterClause": self._filter_clause, u"filters": filter_list}
         for key in output_dict:
             yield key, output_dict[key]
+
+    def __eq__(self, other):
+        if isinstance(other, FilterGroup):
+            return (
+                self.filter_clause == other.filter_clause
+                and self._filter_set == other._filter_set
+            )
+        elif isinstance(other, (tuple, list)):
+            return tuple(self) == tuple(other)
+        elif isinstance(other, string_type):
+            return str(self) == other
+        else:
+            return False
+
+    def __contains__(self, item):
+        return item in self._filter_set