+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The intended audience of this file is for py42 consumers -- as such, changes that don't affect
 how a consumer would use the library (e.g. adding unit tests, updating documentation, etc) are not captured here.
 
-## Unreleased
+## 1.7.0 - 2020-07-21
 
 ### Added
 
@@ -17,6 +17,19 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
     - `sdk.users.get_roles()`
     - `sdk.users.add_role()`
     - `sdk.users.remove_role()`
+
+- `__eq__` and `__hash__` methods to the `py42.sdk.queries.query_filter.QueryFilter` class to enable easier comparison of filters
+- `__eq__` and `__contains__` methods to the `py42.sdk.queries.query_filter.FilterGroup` class to enable easier comparison of and membership tests of filter groups
+
+### Changed
+
+- When calling `__str__` or `__iter__` on a `FilterGroup` instance, the filter results have `set()` called on them to remove duplicate filters (if they exist) as well
+    as sorts the results. This enables comparing two `FilterGroup`s that might have been constructed differently but ultimately return the exact same results in a query.
+- `FilterGroup.filter_clause` property now has a setter, making it easy to change the clause on an existing filter group.
+
+### Removed
+
+- `filter_clause` arg on `FilterGroup.from_dict` method. The clause will automatically be derived from the dict itself.
 
 ## 1.6.2 - 2020-07-10